requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
akshare==1.17.59
pandas>=2.0.0
numpy>=1.24.0
requests>=2.28.0
python-dotenv>=1.0.0
